Prepare for Release
Remove static height option, add name and description
